Implement initial acceptance test
This change adds an errand that checks read availability during a
recreate of an etcd cluster. We chose to go with `bosh recreate` over a
stemcell deploy, as the behavior should be almost identical. Currently
this is implemented as a ginkgo test suite. The read failure tolerance
is hard coded to 0.

[#154321671]

Signed-off-by: Iain Sproat <isproat@pivotal.io>