Implement read-availability-during-network-partition errand
This test suite will create a network partition around a single node in
the cluster. It still allows access to the partitioned node from BOSH,
over SSH, and from the ETCD Client located on the acceptance errand VM.
It then performs a bosh deployment simulation by running `monit restart`
on each of the ETCD nodes. Throughout the experiment, it checks the read
availability of a single key in the ETCD cluster.

[#154321671]

Signed-off-by: Iain Sproat <isproat@pivotal.io>